rename: 로그인/회원가입 관련 api 폴더명 변경 ( user -> auth )
